Adding the plus minus control to change quantity

+++ app/src/main/java/com/piscos/soni/shoppinglist/ShoppingListItem.java
@@ -2,7 +2,7 @@ package com.piscos.soni.shoppinglist;
 
 public class ShoppingListItem extends ProductListItem{
 
-    public double mQuantity;
+    public int mQuantity;
     public ShoppingListItem(){
         mQuantity = 0;
     }
